refactor: move muxrpc setup into index.js, gonna go a different way to implement mounting...
